[DPEDE-409] Change development to date web component, test edge cases
